chore: add comment for jjdescriptionHeader
I always keep looking up what the different multis mean to understand
the jjdescriptionHeader pattern when I read this file.

The comment clarifies what this pattern matches, so the pattern doesn't
have to be understood by those reading the file and wondering whether
they are interested in modifying jjdescriptionHeader or not.

Change-Id: I6a6a6964055daa27d51dca794eb6ac3c454f9f56